Implemented full-featured parallax scrolling, however...
The original parallax implementation (bc3e7459c6c584dea46de2375874c4eb9eae083e) is limited by the intentionally programmed assumption that the next scanline on a cog would have only one line between it and the current scanline where parallaxing could take place. This is because an algorithm hadn't been reasonably developed to identify the next parallaxing scanline index closest to the next scanline (cursl+NUM_REN_COGS). If the next scanline to be rendered on a cog has more than one parallax between it and the current scanline, then you would have to determine which is the most recent - or if there is no most recent and instead the next parallax index in the array is the next time the parallaxing changes. So, it is much simpler to simply assume that the next parallax is either on the same or later scanline than the next scanline to be rendered. With the algorithm implemented in this commit however, an attempt is made to overcome this limitation, resulting in a bug screwing up horizontal wrapping and ONLY horizontal wrapping, not vertical. It appears that when the tile loading routine identifies that it needs to return to the beginning of the tile map, it's behind one row of tiles. Basicaly, whereas nominal behavior would be to start displaying the tiles at the very first row of tiles, it instead starts at the second row of tiles. This results in the last row of tiles rendered after the wrap being basically garbage graphical data. Critical debugging note: NUM_PARALLAX_REGS must be 1 greater than the number of actual parallaxing regions to get any parallaxing at all, even if it's just one region scrolling the entire screen. This at least gives a minimal debugging startpoint where NUM_PARALLAX_REGS = 2, and only one region is attempted to be scrolled.